chore: 增加对 uTools 环境的判断

+++ app/store/chat.ts
@@ -29,7 +29,12 @@ import { ModelConfig, ModelType, useAppConfig } from "./config";
 import { useAccessStore } from "./access";
 import { collectModelsWithDefaultModel } from "../utils/model";
 import { createEmptyMask, Mask } from "./mask";
-import { clearGlobalAsk, setupGlobalAsk } from "../utils/utools";
+import {
+  clearGlobalAsk,
+  isUTools,
+  setupGlobalAsk,
+  storage,
+} from "../utils/utools";
 
 export type ChatMessageTool = {
   id: string;
@@ -758,10 +763,24 @@ export const useChatStore = createPersistStore(
       async clearAllData() {
         await Promise.all([
           indexedDBStorage.clear(),
-          utools.db.promises.allDocs().then((docs) => {
-            docs.forEach((doc) => {
-              utools.db.promises.remove(doc._id);
-            });
+          new Promise((resolve, reject) => {
+            if (isUTools) {
+              resolve(
+                utools.db.promises
+                  .allDocs()
+                  .then((docs) => {
+                    docs.forEach((doc) => {
+                      utools.db.promises.remove(doc._id);
+                    });
+                    return true;
+                  })
+                  .catch(reject),
+              );
+            } else {
+              // @ts-expect-error - Browser Environment
+              storage.clear();
+              resolve(true);
+            }
           }),
         ]);
         location.reload();
